pyproject.toml
[tool.poetry]
name = "paper"
version = "0.1.0"
description = ""
authors = ["jhnnsrs <jhnnsrs@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
jinja2 = ">=3"
pyyaml = "^6.0.1"
pydantic = "<2"
cryptography = "^41.0.3"
namegenerator = "^1.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
